refactor: Email 인증 파일 경로 수정
