+++ src/main/java/tosiltosil/backend/module/auth/domain/response/email/EmailAuthResponse.java
@@ -1,4 +1,4 @@
-package tosiltosil.backend.module.email.domain.response;
+package tosiltosil.backend.module.auth.domain.response.email;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
